Rename `config_base` to `models`
